Refactor the jQuery plugin to its own file
